[chore] #99 module,repository,service,datastore 용도별 분리
