+++ app/src/main/java/com/android/go/sopt/winey/domain/repository/FeedRepository.kt
@@ -1,0 +1,25 @@
+package com.android.go.sopt.winey.domain.repository
+
+import androidx.paging.PagingData
+import com.android.go.sopt.winey.data.model.remote.request.RequestPostLikeDto
+import com.android.go.sopt.winey.data.model.remote.response.ResponsePostWineyFeedDto
+import com.android.go.sopt.winey.domain.entity.Like
+import com.android.go.sopt.winey.domain.entity.WineyFeed
+import kotlinx.coroutines.flow.Flow
+import okhttp3.MultipartBody
+import okhttp3.RequestBody
+
+interface FeedRepository {
+    suspend fun getWineyFeedList(): Flow<PagingData<WineyFeed>>
+
+    suspend fun getMyFeedList(): Flow<PagingData<WineyFeed>>
+
+    suspend fun postFeedLike(feedId: Int, requestPostLikeDto: RequestPostLikeDto): Result<Like>
+
+    suspend fun deleteFeed(feedId: Int): Result<Unit>
+
+    suspend fun postWineyFeed(
+        file: MultipartBody.Part?,
+        requestMap: HashMap<String, RequestBody>
+    ): Result<ResponsePostWineyFeedDto?>
+}
