+++ app/src/main/java/com/android/go/sopt/winey/domain/repository/RecommendRepository.kt
@@ -1,0 +1,7 @@
+package com.android.go.sopt.winey.domain.repository
+
+import com.android.go.sopt.winey.domain.entity.Recommend
+
+interface RecommendRepository {
+    suspend fun getRecommendList(page: Int): Result<List<Recommend>?>
+}
