+++ app/src/main/java/com/android/go/sopt/winey/presentation/main/recommend/RecommendViewModel.kt
@@ -3,7 +3,7 @@ package com.android.go.sopt.winey.presentation.main.recommend
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.android.go.sopt.winey.domain.entity.Recommend
-import com.android.go.sopt.winey.domain.repository.AuthRepository
+import com.android.go.sopt.winey.domain.repository.RecommendRepository
 import com.android.go.sopt.winey.util.view.UiState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -16,9 +16,10 @@ import javax.inject.Inject
 
 @HiltViewModel
 class RecommendViewModel @Inject constructor(
-    private val authRepository: AuthRepository
+    private val recommendRepository: RecommendRepository
 ) : ViewModel() {
-    private val _getRecommendListState = MutableStateFlow<UiState<List<Recommend>?>>(UiState.Loading)
+    private val _getRecommendListState =
+        MutableStateFlow<UiState<List<Recommend>?>>(UiState.Loading)
     val getRecommendListState: StateFlow<UiState<List<Recommend>?>> =
         _getRecommendListState.asStateFlow()
 
@@ -28,7 +29,7 @@ class RecommendViewModel @Inject constructor(
 
     fun getRecommendList() {
         viewModelScope.launch {
-            authRepository.getRecommendList(1)
+            recommendRepository.getRecommendList(1)
                 .onSuccess { response ->
                     _getRecommendListState.value = UiState.Success(response)
                     Timber.e("성공")
